requirements.txt
openai>=1.0.0
anthropic>=0.25.0
pyyaml>=6.0
pandas>=1.5.0
numpy>=1.21.0
tqdm>=4.64.0